feat: Add a readSingleton that returns the result including relational contents.
